refactor: reorganize endpoint discovery
